api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
httpx==0.24.1
python-multipart==0.0.6
supabase==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
twilio==8.1.0
requests==2.31.0
